pyproject.toml
[project]
name = "my-fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]
